+++ opentelemetry-api/src/opentelemetry/util/_importlib_metadata.py
@@ -36,7 +36,7 @@ def _original_entry_points_cached():
     return original_entry_points()
 
 
-def entry_points(**params):
+def entry_points(**params) -> EntryPoints:
     """Replacement for importlib_metadata.entry_points that caches getting all the entry points.
 
     That part can be very slow, and OTel uses this function many times."""